Move structs and func defs used across files to wm_common
