+++ kivy/input/providers/wm_pen.py
@@ -6,11 +6,7 @@ Support for WM_PEN messages (Windows platform)
 __all__ = ('WM_PenProvider', 'WM_Pen')
 
 import os
-from kivy.input.providers.wm_common import (
-    PEN_OR_TOUCH_SIGNATURE, PEN_OR_TOUCH_MASK, GWL_WNDPROC,
-    WM_MOUSEMOVE, WM_LBUTTONUP, WM_LBUTTONDOWN,
-    WM_TABLET_QUERYSYSTEMGESTURE, QUERYSYSTEMGESTURE_WNDPROC,
-    PEN_EVENT_TOUCH_MASK)
+from kivy.input.providers.wm_common import *
 from kivy.input.motionevent import MotionEvent
 
 
@@ -31,47 +27,11 @@ if 'KIVY_DOC' in os.environ:
 
 else:
     from collections import deque
-    from ctypes.wintypes import (ULONG, UINT, WPARAM, LPARAM,
-                                 HANDLE, BOOL)
-    from ctypes import (Structure, windll, byref, c_int16,
-                        c_int, WINFUNCTYPE, POINTER)
+    from ctypes import windll, byref, c_int16, c_int
     from kivy.input.provider import MotionEventProvider
     from kivy.input.factory import MotionEventFactory
 
-    LRESULT = LPARAM
-    WNDPROC = WINFUNCTYPE(LRESULT, HANDLE, UINT, WPARAM, LPARAM)
-
-    class RECT(Structure):
-        _fields_ = [
-            ('left', ULONG),
-            ('top', ULONG),
-            ('right', ULONG),
-            ('bottom', ULONG)]
-
-        x = property(lambda self: self.left)
-        y = property(lambda self: self.top)
-        w = property(lambda self: self.right - self.left)
-        h = property(lambda self: self.bottom - self.top)
     win_rect = RECT()
-
-    try:
-        windll.user32.SetWindowLongPtrW.restype = WNDPROC
-        windll.user32.SetWindowLongPtrW.argtypes = [HANDLE, c_int, WNDPROC]
-        SetWindowLong_wrapper = windll.user32.SetWindowLongPtrW
-    except AttributeError:
-        windll.user32.SetWindowLongW.restype = WNDPROC
-        windll.user32.SetWindowLongW.argtypes = [HANDLE, c_int, WNDPROC]
-        SetWindowLong_wrapper = windll.user32.SetWindowLongW
-
-    windll.user32.GetMessageExtraInfo.restype = LPARAM
-    windll.user32.GetMessageExtraInfo.argtypes = []
-    windll.user32.GetClientRect.restype = BOOL
-    windll.user32.GetClientRect.argtypes = [HANDLE, POINTER(RECT)]
-    windll.user32.CallWindowProcW.restype = LRESULT
-    windll.user32.CallWindowProcW.argtypes = [WNDPROC, HANDLE, UINT, WPARAM,
-                                              LPARAM]
-    windll.user32.GetActiveWindow.restype = HANDLE
-    windll.user32.GetActiveWindow.argtypes = []
 
     class WM_PenProvider(MotionEventProvider):
